Add contents filtering transfer decorator
It makes it possible to set a remote repository proxy to not provide
releases/snapshots if needed. It filters the maven-metadata.xml contents
and also listing of a GAV. However it does not affect normal directory
listing which does not contain any artifacts, i.e. GA listing still
contains all version directories, only their listing gets filtered.

Because of all the filtering has to be applied to remote locations only
it had to be placed in httpclient to have HttpLocation interface visible.

WHen testing I realized that a remote listing gets composed from locally
cached listing, which is empty, and the remote listing obtained via
HttpListing. However the remote listing did not get processed via transfer
decorator, so the resulting listing was not filtered. I added the manual
decorator call to not duplicate the listing filtering logic in the
decorator itself.