Fix listing to not contain links not leading into subdirectories
When there is a chain of multiple Indy instances, then the listing have
grown bigger by the links from the footer saying what were the source
locations for a group content. For example "builds-untested" or "http:/"
(this one is really weird).

The fix checks if the link points to the same server into a subdirectory
of the currently listed directory and includes only the links that do.