Refactor TransferDecorator pipelining to allow pipelining when ctors aren't available
In some cases we may instantiate TransferDecorators via CDI producer methods. When this
happens, we don't necessarily have the perfect knowledge of the whole system needed to
create the decorator pipeline using the ctor-based nesting we had before.

This refactor creates a TransferDecoratorPipeline class, and reverts AbstractTransferDecorator
to a simple (empty) decorator implementation. The TransferDecoratorPipeline accepts either
a varargs array or list of TransferDecorators to invoke, then uses list iteration to
handle pipeline execution.