Refactor ChecksummingTransferDecorator and friends to produce TransferMetadata instances
We need to do more with calculated checksums than simply write them to files. We need
the ability to specify a consumer instance to accept the calculated checksums + content length
resulting from a read or write operation.

This refactor migrates some classes from Indy used to organize checksums + content length into
a domain object, namely TransferMetadata (and associated ContentDigest enum). If a
TransferMetadataConsumer instance is passed into the ChecksummingTransferDecorator, these
in-memory representations of checksums will be passed, along with the Transfer they describe,
to the consumer. This allows us to nominate a component to "listen" to checksums as they are
calculated, in addition to (or instead of) writing to checksum files.

Checksum files can still be written, and whether they are written depends on the TransferOperation
set passed into the ChecksummingTransferDecorator. All read operations are assumed to use
TransferOperator.DOWNLOAD.