Introduce ability to force checksumming and delete checksums from consumer.
In some cases we need to be able to override the way ChecksummingTransferDecorator
is initialized, to force it to calculate checksums on a read, for instance, when
reader checksumming isn't enabled.

We can use EventMetadata for this. By setting a flag in the EventMetadata we can
override the base configuration of the decorator.