Add needsMetadataFor method for checksum avoidance
We have the ability to configure ChecksummingTransferDecorator to be
over-eager in calculating checksums. For example, when doing successive
reads of a file and when read-checksumming is enabled, we will constantly
be calculating checksums that we already have cached, without anything
changing.

This change adds a needsMetadataFor method to TransferMetadataConsumer,
which can allow the ChecksummingTransferDecorator to avoid calculating
checksums on file read / write if the consumer already has it cached.

Cases where this has NO effect:

* When the TransferOperation matches the list of operations for which we
want to write checksum files (we'll always rewrite these files when these
operations happen)

* When we specify FORCE_CHECKSUM == Boolean.TRUE in EventMetadata

This could probably use some more efficiency improvement around creation
of checksum files (or avoiding that when they already exist), but it
should help in a lot of common cases.