Extending @ligangty's PR#184 to use a logger name specific to content deletion
This uses a separate logger, with the 'org.commonjava' base namespace, an
intermediate namespace of 'content' (for any content action), and a
specific name of 'delete'. This will allow us to isolate these messages
from others and push them to a different log appender. We should also
be able to reuse the name for any location where content is being deleted
in Commonjava libraries / applications.