Refactor ChecksummingTransferDecorator to accept two ChecksummingDecoratorAdvisors, one for read, one for write
There isn't enough control over when checksums get calculated, and the controls over when checksum files
get written are confusing. The result is that Indy is calculating checksums on literally everything when
it only needs to do so in specific cases when Folo tracking is enabled.

This change provides a new interface, ChecksummingDecoratorAdvisor, which allows Indy to inject specific
logic about when checksums should be calculated, stored, or skipped. The old constructor is still present
but deprecated, and the old ctor params are used to create two internal advisors that replicate the old
behavior. ChecksummingDecoratorAdvisor returns ChecksumAdvice, which is an enum that has three values:
NO_DEORATE, CALCULATE_NO_WRITE, and CALCULATE_AND_WRITE. This covers the three branches of the logic
in the decorator now. All AbstractChecksumGenerator implementations now accept a writeChecksumFiles
constructor parameter, to control whether the retrieve and lock Transfers corresponding to their
checksum file type.

The change also provides DisabledChecksummingDecoratorAdvisor, which can be used to completely disable
non-forced checksumming for reads or writes, if that's appropriate.