Change the way of ISPN path lock to avoid duplicate transaction start
   We need to think about the way of the ISPN lock and wait. If directly
use the cache.lock but not consider if the lock has been acquired by another
thread, the ISPN lock will fail with a RuntimeException. So we need to let the
thread wait for the ISPN lock until it's released by the thread holds it. It's
like "tryLock" and "wait" way of ReentrantLock.