NOS-940: directory listings with no trailing slash
When a directory listing is requested without the trailing slash in the
url, it gets cached as a file and subsequent requests for files under
this directory leads to exception saying that the directory could not be
created because a file with the same name already exists.

This fix checks if it is the case and if it finds a file in the path, it
deletes it to allow the directory creation.

This applies only for http downloads. Any other operations don't perform
any deletions. It is also possible to create an http download with the
deletion switched off, but this is not used ATM.