Throw TransferLocationException for wider range of 4xx / 5xx status codes
Skipping 404 and 502-504, which are temporary problems, Galley's HTTP transfer
jobs should throw TransferLocationException for other 4xx/5xx, since they
represent either a server-side problem that needs to be corrected (not temporary),
or a misconfiguration of the way the Location is setup to talk to the server.

This will trigger Indy (for example) to disable the repository associated with
the error.